Move Issue license page to Primary folder.
