+++ client/jetpack-cloud/sections/partner-portal/primary/issue-license/index.tsx
@@ -18,7 +18,7 @@ import LayoutTop from 'calypso/jetpack-cloud/components/layout/top';
 import PartnerPortalSidebarNavigation from 'calypso/jetpack-cloud/sections/partner-portal/sidebar-navigation';
 import { useDispatch } from 'calypso/state';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
-import AssignLicenseStepProgress from '../assign-license-step-progress';
+import AssignLicenseStepProgress from '../../assign-license-step-progress';
 import IssueLicenseContext from './context';
 import { useProductBundleSize } from './hooks/use-product-bundle-size';
 import useSubmitForm from './hooks/use-submit-form';
@@ -26,7 +26,7 @@ import LicensesForm from './licenses-form';
 import ReviewLicenses from './review-licenses';
 import TotalCost from './total-cost';
 import type { SelectedLicenseProp } from './types';
-import type { AssignLicenceProps } from '../types';
+import type { AssignLicenceProps } from '../../types';
 
 import './style.scss';
