+++ client/jetpack-cloud/sections/partner-portal/primary/issue-license/review-licenses/pricing-breakdown.tsx
@@ -1,7 +1,7 @@
 import formatCurrency from '@automattic/format-currency';
 import { Icon, check } from '@wordpress/icons';
 import { useTranslate } from 'i18n-calypso';
-import { useLicenseLightboxData } from '../../license-lightbox/hooks/use-license-lightbox-data';
+import { useLicenseLightboxData } from '../../../license-lightbox/hooks/use-license-lightbox-data';
 import { getProductPricingInfo, getTotalInvoiceValue } from '../lib/pricing';
 import type { SelectedLicenseProp } from '../types';
 import type { SiteDetails } from '@automattic/data-stores';
