+++ client/jetpack-cloud/sections/partner-portal/primary/product-price-with-discount-info/index.tsx
@@ -4,7 +4,7 @@ import { useTranslate } from 'i18n-calypso';
 import { useSelector } from 'calypso/state';
 import { APIProductFamilyProduct } from 'calypso/state/partner-portal/types';
 import { getProductsList } from 'calypso/state/products-list/selectors';
-import { getProductPricingInfo } from '../../issue-license-v2/lib/pricing';
+import { getProductPricingInfo } from '../issue-license/lib/pricing';
 
 import './style.scss';
